% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_err.R
\name{std_err}
\alias{std_err}
\title{Calculate standard error}
\usage{
std_err(x)
}
\arguments{
\item{x}{A vector of numerical data.}
}
\value{
A numeric value.
}
\description{
Using a vector of samples, calculate the standard error of the
mean.
}
\examples{
    std_err(x = c(4, 7, 2, 7, 5, 9))

}
